Improve fp16 numerical stability: add gradient clipping, conservative loss scaling, and adaptive loss scaling

+++ launch_experiment.py
@@ -154,7 +154,7 @@ def launch_experiment(args):
     cmd = [
         'accelerate', 'launch',
         '--multi_gpu',
-        '--mixed_precision', 'bf16',
+        '--mixed_precision', 'fp16',
         '--num_processes', static_config['num_process'],
         '--main_process_port', static_config['master_port'],
         'run_main.py',
